requirements.txt
requests
beautifulsoup4
gspread
oauth2client
tqdm
python-dotenv
playwright
tenacity
urllib3
